Refactor flow viewer PR4 composition
