+++ src/ui-component/editor/ajsFlow/nodes/ConditionNode.tsx
@@ -10,7 +10,7 @@ import {
   AjsNode,
   buildNodeSxProps,
   nodeActionsSxProps,
-  NameOrComment,
+  NodeNameAndComment,
   TyTitle,
 } from "./AjsNode";
 import {
@@ -66,8 +66,7 @@ const ConditionNode: FC<ConditionNodeProps> = ({
           )}
         </Box>
       </Stack>
-      <NameOrComment value={label} />
-      <NameOrComment value={comment} />
+      <NodeNameAndComment label={label} comment={comment} />
     </>
   );
 };
